refactor and downgrade to sdk 4.1.1
